backend/requirements.txt
crewai>=0.28.0
langchain>=0.0.335
openai>=1.3.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
